Map position and Real position, Cardinal point
IMPORTANT : I changed the function to get the direction of the robot
from 'get_d' to 'get_w', and this function returns now a double from 0
to 360 degrees.
The functions 'get_x' and 'get_y' now return the values encoded with the
map resolution. Furthermore I added several functions to get the real
values for X and Y (in mm) called 'get_realX' and 'get_realY', and a
function to get the cardinal point of the direction of the robot. For
more, read comments in 'localization.h'.