Localization sent the CellCenter event only once by cell
(+ a minor fix in init_localization)
This changes also the MainController behaviour at initialization. The
localization begins also later, for the mapping to have the time to
initialize values for the first cells. When localization starts, it
sends the CellCenter event to the MainController to wake it up, the
MainController initialize the localization and if the robot is already
in the center, decides what to do next. Otherwise enter in the main loop
and wait for the next CellCenter event.