Change the behaviour of the localization resource
The localization resource, called UpdateLocker before, now called
SyncLocalization, was defined to make sure that a Task which want to get
x and y values for example will always be sure to have synchronised
values. But it was not the case with the previous implementation, and it
slowed the program for nothing. Now it's better, and the resource is get
only when needed.